requirements.txt
# Core Frameworks
crewai==0.28.8
langgraph==0.0.69
langchain==0.1.13
langchain-openai==0.0.8

# LLM Providers
openai==1.14.2
anthropic==0.21.3

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
rembg==2.0.54  # Background removal
numpy==1.26.4

# HTTP Client
requests==2.31.0
aiohttp==3.9.3

# Web Framework (for API)
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# Data Validation
pydantic==2.6.4
pydantic-settings==2.2.1

# Storage
minio==7.2.5
redis==5.0.3

# Task Queue
celery==5.3.6

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.2

# Monitoring (Optional)
langsmith==0.1.31
